camino BE/requirements.txt
antlr4-python3-runtime==4.7.2
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.3
certifi==2017.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
colorlover==0.3.0
contextlib2==0.5.5
cufflinks==0.17.3
decorator==4.4.2
defusedxml==0.6.0
Django==2.0
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.3
easypost==4.0.0
entrypoints==0.3
Flask==1.1.1
httpie==0.9.9
idna==2.6
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.13.2
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
Markdown==2.6.9
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==6.0.3
npm==0.1.1
numpy==1.13.3
oauthlib==3.1.0
olefile==0.44
optional-django==0.1.0
packaging==20.1
pandas==1.0.2
pandocfilters==1.4.2
parso==0.6.2
pdfkit==0.6.1
pickleshare==0.7.5
Pillow==4.3.0
plotly==4.5.4
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psycopg2==2.7.3.2
py==1.8.1
Pygments==2.2.0
pyparsing==2.4.6
PyQt5==5.11.2
PyQt5-sip==4.19.12
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.5
python-dateutil==2.8.1
pytz==2017.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
requests==2.18.4
requests-oauthlib==1.3.0
retrying==1.3.3
schema==0.7.1
scikit-learn==0.22.2.post1
scipy==1.3.1
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.14.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
tornado==6.0.4
traitlets==4.3.3
tweepy==3.8.0
urllib3==1.22
virtualenv==15.1.0
virtualenvwrapper-win==1.2.6
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wkhtmltopdf==0.2
zipp==3.0.0
